+++ src/core/ui/shapes.ts
@@ -1,5 +1,5 @@
 /// <reference path="./../../reference.d.ts" />
-namespace Lightening.UI {
+namespace Lightning.UI {
     export namespace Shapes {
         /**
          * @description Draw a square