+++ src/core/ui/sprite.ts
@@ -1,12 +1,26 @@
 /// <reference path="./../../reference.d.ts" />
 
-namespace Lightening.UI {
+namespace Lightning.UI {
     export class Sprite extends PIXI.Sprite {
         
         private _body;
 
         constructor(texture:PIXI.Texture = null) {
             super(texture);
+        }
+
+        enableBody(val:boolean) {
+            if(val) {
+                
+            }
+        }
+
+        setAnchor(aX, aY = null):void {
+            if(!aY) {
+                this.anchor = new PIXI.Point(aX, aX);
+            } else {
+                this.anchor = new PIXI.Point(aX, aY);
+            }
         }
 
         set body(body) {