+++ typings/globals/pixi.js/index.d.ts
@@ -1209,8 +1209,8 @@ declare module PIXI {
 
         constructor(texture?: Texture);
 
-        protected _anchor: ObservablePoint;
-        anchor: ObservablePoint;
+        protected _anchor: Point;
+        anchor: Point;
         protected _texture: Texture;
         protected _transformTrimmedID: number;
         protected _textureTrimmedID: number;